Added feature: Send cpu allocation to PDCP in double format

+++ manager_msg.h
@@ -149,7 +149,7 @@ typedef struct
 {
 	int 		module_id;     							/*!< \brief  Virtualized module identifier      */
 	int  		resource_req;      						/*!< \brief  Amount of CPU resource request from an module */
-	int			resource_rsp;							/*!< \brief  Amount of allocated CPU resource from the cloud manager */
+	double		resource_rsp;							/*!< \brief  Amount of allocated CPU resource from the cloud manager */
 	double		down_BW_req;							/*!< \brief  Downlink Bandwidth request */
 	double		down_BW_rsp;							/*!< \brief  Downlink Bandwidth allocation from CM */
 	double		up_BW_req;								/*!< \brief  Uplink Bandwidth request */
